chore: split copy_with_rewrites into several files
