+++ packages/arpa_integration_tools/util.ts
@@ -1,0 +1,14 @@
+import * as path from "path";
+import * as fs from "fs/promises";
+
+export function truncateExtension(filePath: string): string {
+  const { dir, name } = path.parse(filePath);
+  return path.join(dir, name);
+}
+
+export function exists(fpath: string): Promise<boolean> {
+  return fs.access(fpath).then(
+    () => true,
+    () => false
+  );
+}
